fix: float步号支持 + lyric_scene_designer JSON解析调试
- pipeline_loader.py: int() → float() 支持 7.5/9.1 等小数步号
- lyric_scene_designer/impl.py: 添加 raw 返回保存 + 详细错误信息
- 验证: 7.5-9.0 全链路通过, 32个歌词场景用7种视觉类型, 零大字报

+++ pipeline_loader.py
@@ -299,8 +299,8 @@ def run_pipeline(
 
     # 解析步骤范围
     parts = steps.split("-")
-    start_step = int(parts[0])
-    end_step = int(parts[1]) if len(parts) > 1 else start_step
+    start_step = float(parts[0])
+    end_step = float(parts[1]) if len(parts) > 1 else start_step
 
     # 视频分辨率
     video_width, video_height = (1080, 1920) if vertical else (1920, 1080)
